requirements.txt
numpy==1.16.2
pandas==0.24.2
requests==2.21.0
matplotlib==3.0.3
Pillow==6.0.0
